+++ src/Matchers/MatcherInterface.php
@@ -13,5 +13,5 @@ interface MatcherInterface
      *
      * @return bool
      */
-    public function match(ServerRequestInterface $request): bool;
+    public function __invoke(ServerRequestInterface $request): bool;
 }