+++ src/Matchers/Path.php
@@ -25,7 +25,7 @@ class Path implements MatcherInterface
     /**
      * {@inheritdoc}
      */
-    public function match(ServerRequestInterface $request): bool
+    public function __invoke(ServerRequestInterface $request): bool
     {
         $path = $request->getUri()->getPath();