+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.5.0 - 2017-09-22
+
+### Changed
+
+* Changed `MatcherInterface` to allow use callables.
+* Updated to `http-interop/http-middleware#0.5`
+* Updated phpunit to `^6.0`
+
 ## 0.4.0 - 2017-02-29
 
 ### Changed